Move search forms out of ProteinService
